Initial steps for the password configs

+++ backend/src/main/java/br/uema/ppg/IntelligentAssistent/setores/SetorService.java
@@ -31,6 +31,9 @@ public class SetorService {
     }
 
     public void deletarSetor(Long id){
+        if (!setorRepository.existsById(id)){
+            throw new RuntimeException("Setor não criado ou encontrado com o id: " + id);
+        }
         Setor salvo = buscarPorId(id);
         setorRepository.delete(salvo);
     }
